+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Pandashark's Portfolio の大まかな機能変化記録
 
+## 2022-04-23
+
+### Product Added
+
+- Simple Kanban Planner を追加
+
 ## 2022-03-13
 
 ### Development Added